+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Path field to retrieve of path using fsspec
+
 ## [0.7.1] - 2024-10-22
 ### Added
 - Option to add unit labels by passing a mapping in QuantityField unit_options